Added Cyclic and SimContext EClasses.
This change allows SimMachineDef and SimOperationDef to inherit the same
attributes of SimContext: inputContext and outputContext.

It also allows SimModule, SimControllerDef, SimMachineDef and
SimOperationDef to inherit the same attributes of Cyclic: cycleDef.

This simplifies the methods by allowing polymorphism over objects of
type Cyclic and SimContext, where SimContext is a Context as well (so it
includes required, provided and defined interfaces, variables and events
as necessary.