Version of the RoboSim metamodel targeting revision of the role of 'exec'.
* Made Cyclic and SimContext abstract classes.

This isn't a genuine revision of the metamodel per see, but introduces a
custom implementation of the ExecTrigger (Java) class that when
instantiated via the RoboSimFactoryImplCustom class attempts to find the
'exec' event in scope and sets it as the value of the (now) mandatory
event attribute. This is a way to ensure that upon use of the class
ExecTrigger later by the Xtext framework in the robosim-textual plug-in
that no ExecTrigger is constructed in an ill-way. With this approach we
also guarantee that the textual language, and the constructs of the
metamodel remain the same! Thus, for example, no type related changes
are required to the CSP generator.

Further testing is required with the model-to-model work with Epsilon to
ensure this does not introduce unexpected problems.