Eliminated custom implementation of ExecTrigger.
As it turns out this was a rather arcane and proved problematic for
serialization of RoboSim/C2S resources that used ExecTrigger. Therefore
the metamodel in this version is completely compatible with the existing
deployed version.

Instead, I've found a way to make Xtext accept 'exec' as both a keyword,
and a event reference, which is a cleaner solution.